requirements.txt
aiohttp>=3.6
appdirs>=1.4.0
cssselect>=1.1.0
lxml>=4.0.0
mutagen>=1.31
pillow>=8.0.0
tqdm>=4.28.1
unidecode>=1.1.1
web_cache>=1.1.0
